stupid fix

+++ internal/api/endpoint_stat.go
@@ -175,7 +175,7 @@ func ChSelectMetricDuration(duration time.Duration, metricID int32, user, table,
 			5: ok,
 			6: getStatTokenName(user),
 			7: user,
-			8: strconv.Itoa(int(metricID % 16)), // experimental to see load distribution if we shard data by metricID
+			8: strconv.Itoa(int(uint32(metricID) % 16)), // experimental to see load distribution if we shard data by metricID
 		},
 	).Value(duration.Seconds())
 }
